refactor(runtime): move server entrypoint to src with root shim
