+++ test/unit/validation.test.mjs
@@ -18,17 +18,17 @@ describe("package.json files allowlist", () => {
     }
   });
 
-  test("every local JS module imported by index.js is in the files allowlist", () => {
-    const indexSrc = fs.readFileSync(path.join(root, "index.js"), "utf8");
+  test("every local JS module imported by src/index.js is in the files allowlist", () => {
+    const entrypointSrc = fs.readFileSync(path.join(root, "src", "index.js"), "utf8");
     const localImports = [
-      ...indexSrc.matchAll(/^\s*import\b(?:[\s\S]*?\bfrom\s*)?["'](\.\/[^"']+)["']/gm),
-    ].map((m) => m[1].replace(/^\.\//, ""));
+      ...entrypointSrc.matchAll(/^\s*import\b(?:[\s\S]*?\bfrom\s*)?["'](\.\.\/[^"']+)["']/gm),
+    ].map((m) => m[1].replace(/^\.\.\//, ""));
 
     for (const file of localImports) {
       const covered =
         allowlist.includes(file) ||
         allowlist.some(entry => entry.endsWith("/") && file.startsWith(entry));
-      assert.ok(covered, `"${file}" is imported by index.js but missing from package.json files allowlist`);
+      assert.ok(covered, `"${file}" is imported by src/index.js but missing from package.json files allowlist`);
     }
   });
